#399 Moviendo lógica de submenues al controller en el paquete
